refactor(castore/directorysvc/combinators): cleanup
Express the whole populating of the Directory graph as a try_fold, and
rename variables a bit to make it more meaningful. We introduce a root()
helper function to ValidatedDirectoryGraph, which we'll use to get the
root, rather than storing it in a let binding.

Change-Id: I4f942c6484dbafe7957d9ef85f205d632155a167
Reviewed-on: https://cl.snix.dev/c/snix/+/30745
Tested-by: besadii
Reviewed-by: edef <edef@edef.eu>
Autosubmit: Florian Klink <flokli@flokli.de>