refactor(snix/castore): drop MemoryDirectoryService, add redb+memory
We don't need multiple in-memory variants, and the naive HashMap one
doesn't support transactions.

Redb has an in-memory mode, so witch to that one. Update the URL scheme,
use `redb+memory:`, instead of the previous `redb://` without a path.

Also start being a bit stricter w.r.t. authority in the URI.

Make RedbDirectoryServiceConfig only accept a Option<PathBuf>.
For now it's enough to model whether it's in-memory or not.

Drop support for `redb:` URLs without a path, and introduce
`redb+memory:` (without authority and path) for the in-memory variant.

Change-Id: I0d7dde4348d33722f70eaf11f95cf27e7999afdc
Reviewed-on: https://cl.snix.dev/c/snix/+/30778
Tested-by: besadii
Reviewed-by: Connor Brewster <cbrewster@hey.com>