feat(castore/directorysvc/redb): support read-only mode
redb 3.0.0 introduced read-only multi-process support.
If read_only is passed as a parameter to redb (either via store
composition or as a URL), the database is opened in read-only mode,
allowing multiple processes to access at the same time (as long as all
of them open in r/o mode).

Opening the database on one mode or the other returns a different type,
so we use an enum to store one or the other type of the struct.

Change-Id: Ie0f0f90ccc6859d1be766e88e84d538a4a2f6384
Reviewed-on: https://cl.snix.dev/c/snix/+/30779
Reviewed-by: Connor Brewster <cbrewster@hey.com>
Reviewed-by: edef <edef@edef.eu>
Tested-by: besadii