refactor(castore/DirectoryGraph): make drain order explicit
Make the order explicit, by the function name, and DirectoryOrder
private.

Most consumers use the Leaves-To-Root drain order, or yield items when
inserting in Root-To-Leaves order succeeded, but not all.

Having explicit functions makes this easier to track, and we don't want
to decide about drain order at runtime, so we don't need the enum.

Change-Id: Ia63f905089f710d299262a2920a18b237521d288
Reviewed-on: https://cl.snix.dev/c/snix/+/30801
Tested-by: besadii
Autosubmit: Florian Klink <flokli@flokli.de>
Reviewed-by: edef <edef@edef.eu>