feat(snix/store/pathinfo): add PathInfoService::has
Allow asking if a PathInfo is stored without actually fetching it.

Some databases (not redb) are able to respond to this request faster.
In the case of the NixHTTPPathInfoService, we can avoid downloading
entirely, if it's just a HEAD request.

Unfortunately Nix doesn't send HEAD requests (anymore?), so this doesn't
really help us with https://git.snix.dev/snix/snix/issues/59.

Change-Id: I712b3791d432258a184f848a6c64ecffd53dc7ce
Reviewed-on: https://cl.snix.dev/c/snix/+/30816
Reviewed-by: Yureka <snix@yuka.dev>
Tested-by: besadii