refactor(castore/directorysvc): have RTL validator ask for root digest only
This is similar to the constructor of the graph variant in cl/30795.

Only ask for the expected root digest, not require to initialize with
the first directory.

This makes the consumer side much easier, and less prone to mistakes.
Previously, the first element would need to be fetched separately, and
it was easy to construct the validator without checking the root digest
to be what as passed in the request.

Now there's no need to treat the first element differently, so it's much
harder to actually have another digest in scope.

Change-Id: I8ada266e531381421ff752eaf16a099ef5988a36
Reviewed-on: https://cl.snix.dev/c/snix/+/30825
Tested-by: besadii
Autosubmit: Florian Klink <flokli@flokli.de>
Reviewed-by: Brian Olsen <brian@maven-group.org>