refactor(snix): deal with tracing args consistently
Add a TracingBuilder::handle_tracing_args() helper, which takes a
populated snix_tracing::TracingArgs.

In there, we define the --otlp option. This ensures we have consistent
behaviour across the different CLI entrypoints.

Also update castore-http to make use of it. It previously did set up
tracing on its own, without any OTLP setup.

Setting up tracing needs to happen inside a tokio runtime context, so we
now do this inside an async block passed to `tokio_runtime.block_on`.
We also do init_io_handle in there, which is now async, and avoids
unnecessarily passing around the runtime here. It now returns a
SnixStoreIO without the surrounding Rc, as we can deal with how to wrap
it on the callsites.

For now, as the batch processor is super spammy when no collector is
runnig, disable otlp. We need to revisit this once actually working
on metrics, and maybe make it more granular.

Change-Id: I023986215157b9ef2f42edd108a1c5405491b28d
Reviewed-on: https://cl.snix.dev/c/snix/+/30844
Reviewed-by: Brian Olsen <brian@maven-group.org>
Autosubmit: Florian Klink <flokli@flokli.de>
Tested-by: besadii