feat(snix): Use cargo docs-rs and crane for rustdocs
This switches to using cargo docs-rs when building the rustdocs which is
a tool that runs rustdoc command in the same way as docs.rs does when
they generate documentation.

This also means that we stop documenting 3rd-party crates and instead
link to docs.rs for those.

The advantage to using that tool is that we can document which feature
is required for a given module, type or function as well as being able
to use `#[cfg(docsrs)]` to have code and documentation that are only
relevant for rustdoc. It also gives us access to unstable documentation
features.

This requires a nightly rust since it uses unstable features and getting
that setup with an updated rust-overlay and crane seemed the easiest.
And so this commit also does that.

Change-Id: I3917f84d0d38183a9053fa6bbac0df429cd885f0
Reviewed-on: https://cl.snix.dev/c/snix/+/30846
Tested-by: besadii
Autosubmit: Brian Olsen <brian@maven-group.org>
Reviewed-by: Florian Klink <flokli@flokli.de>