refactor(snix/tracing): stop passing around service name
This didn't really work for a while, as can be seen in the
nixos.snix.store grafama, traces get sent with `unknown_service` as
a name.

The documented way is to use the OTEL_SERVICE_NAME env var, so let's do
that as well and simplify this to a boolean for now.

Change-Id: Ic9c806c55e31b512c7249f3eab3335307c2feae5
Reviewed-on: https://cl.snix.dev/c/snix/+/30854
Autosubmit: Florian Klink <flokli@flokli.de>
Tested-by: besadii
Reviewed-by: Brian Olsen <brian@maven-group.org>