fix(snix/nar-bridge): fix shutdown behaviour
Use .with_graceful_shutdown and pass in a future waiting on both
ctrl-c and sigterm.

Once the server terminates, all databases should already be dropped, as
they were only kept alive in the AppState struct, which was moved into
the now terminated server.

This will cause nar-bridge to not have to repair its DB on startup,
improving startup time.

Fixes #56, but we need to deal with the other services separately.

Change-Id: I95f629b66d696c242a94091c36b889f1bdb1f3df
Reviewed-on: https://cl.snix.dev/c/snix/+/30869
Tested-by: besadii
Reviewed-by: Brian Olsen <brian@maven-group.org>
Autosubmit: Florian Klink <flokli@flokli.de>