fix(snix): Fix contents of target/ causing build errors
The `mkDummySrc` in crane which is used by `buildDepsOnly` doesn't
convert the `src` argument to a string to get the `outPath` from
filtered sources but instead uses the original path directly.
It then looks for all files called `Cargo.toml` and puts them in store
in a modified version to make a dummy src that can be used to build deps
only. But because `trybuild` used for ui tests in
`nix-compat-derive-tests` makes a `Cargo.toml` for those tests in
`target/` with absolute paths that gets sucked up by crane and the build
fails.

The solution is to force crane to use the filtered `src` which this
commit does.

Change-Id: If7dd7271f914e838d860c0d576ebe6bcbd731115
Reviewed-on: https://cl.snix.dev/c/snix/+/30886
Autosubmit: Brian Olsen <brian@maven-group.org>
Reviewed-by: Florian Klink <flokli@flokli.de>
Tested-by: besadii