refactor(snix/cli): Move snix binary to snix/cli-eval and rename it
In preparation for implementing #185 the current `snix` binary needed to
be moved out of the way and renamed to something else. Right now it has
been renamed `snix-eval` and is moved to `snix/cli-eval` but in time
it should be moved to `snix/cli/eval` and be called as `snix eval`.

Change-Id: Iba3ad2b44dbda8703f95fae226284d8f54deaf14
Reviewed-on: https://cl.snix.dev/c/snix/+/30895
Autosubmit: Brian Olsen <brian@maven-group.org>
Reviewed-by: Florian Klink <flokli@flokli.de>
Tested-by: besadii