feat(snix/glue): reject __json in builtins.derivation on non-structured attrs
Nix started rejecting this in
https://github.com/NixOS/nix/blob/8093df5255572bde31f16287feabb155a218004e/doc/manual/source/release-notes/rl-2.30.md?plain=1#L13,
it's probably a good idea to do the same.

It's still possible to set `__json` iff `__structuredAttrs` is enabled;
in that case it'll get added as a key /underneath/ the `__json` key
holding the JSON. We can verify this to be consistent with Nix by
comparing the drvPath.

Also move `__json` to a constant.

Link: https://git.snix.dev/snix/snix/issues/33#issuecomment-739
Change-Id: If6012a6ee341caf72c30c54fe2a7f4d5cac58397
Reviewed-on: https://cl.snix.dev/c/snix/+/30901
Reviewed-by: Ryan Lahfa <ryan@lahfa.xyz>
Tested-by: besadii
Autosubmit: Florian Klink <flokli@flokli.de>