fix(web): Update docs to reflect new CLI layout
This updates the documentation to reflect that we now have one
entrypoint for the CLI with several sub-commands instead of the before
where we had several binaries.

This also updates build links to point to the default combined CLI
derivation in `snix/cli/default-cli.nix`.

Change-Id: I2abe67fda40884a3d81cbe128e370655d4a286e4
Reviewed-on: https://cl.snix.dev/c/snix/+/30912
Reviewed-by: Florian Klink <flokli@flokli.de>
Tested-by: besadii
Autosubmit: Brian Olsen <brian@maven-group.org>